+++ spec/offset_request_spec.rb
@@ -4,6 +4,27 @@ require_relative '../lib/tusc/offset_request'
 
 WebMock.disable_net_connect!
 
+RSpec.describe 'TusClient::OffsetRequest (constructor)' do
+  it 'converts upload_url to upload_uri (as URI)' do
+    request = TusClient::OffsetRequest.new(upload_url: 'https://offset.example.com')
+    expect(request.upload_uri).to be_a(URI)
+    expect(request.upload_uri.to_s).to eql('https://offset.example.com')
+  end
+
+  it 'accepts a URI' do
+    upload_uri = URI.parse('https://offset.example.com')
+    request = TusClient::OffsetRequest.new(upload_url: upload_uri)
+    expect(request.upload_uri).to be_a(URI)
+    expect(request.upload_uri.to_s).to eql('https://offset.example.com')
+  end
+
+  it 'requires a valid upload_url' do
+    expect do
+      TusClient::OffsetRequest.new(upload_url: 'invalid_url')
+    end.to raise_error(URI::InvalidURIError, /host/)
+  end
+end
+
 RSpec.describe TusClient::OffsetRequest do
   subject(:request) do
     TusClient::OffsetRequest.new(
@@ -22,12 +43,11 @@ RSpec.describe TusClient::OffsetRequest do
   describe '#perform' do
     let(:expected_offset) { -1 }
     before(:each) do
-
-      stub_request(:head, upload_url).
-        to_return(
+      stub_request(:head, upload_url)
+        .to_return(
           headers: {
             'Upload-Offset': expected_offset.to_s,
-            'Tus-Resumable': '1.0.0',
+            'Tus-Resumable': '1.0.0'
           }
         )
     end